update 1/6/2020
in this Update I handle the exception of an array doen't have the lenght 9x9, and also I create an exception if the element are out of the range 1 to 9.
I try to create each function seperatly that will check the rows, columns and subgrid. and finaly i remove the object that will be inisialized each time